requirements.txt
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
editdistance==0.5.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-core==1.7.2
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.1
nltk==3.6.2
numpy==1.21.1
packaging==21.0
pandas==1.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
regex==2021.7.6
requests==2.26.0
rsa==4.7.2
six==1.16.0
tqdm==4.61.2
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
